soft delete structure and settings added

+++ App.Web.Mvc/Areas/Admin/Controllers/AdminContactController.cs
@@ -1,13 +1,14 @@
 ﻿using App.Web.Data.Concrete;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-
+using System;
+using System.Linq;
 
 namespace App.Web.Mvc.Areas.Admin.Controllers
 {
     [Area("Admin")]
     [Authorize(Policy = "ModeratorPolicy")]
-	public class AdminContactController : Controller
+    public class AdminContactController : Controller
     {
         private readonly AppDbContext _context;
 
@@ -18,8 +19,14 @@ namespace App.Web.Mvc.Areas.Admin.Controllers
 
         public IActionResult Index()
         {
-            var contactMessages = _context.ContactMessages.ToList();
+            var contactMessages = _context.ContactMessages.Where(x => x.DeletedAt == null).ToList();
             return View(contactMessages);
+        }
+
+        public IActionResult DeletedMessages()
+        {
+            var deletedMessages = _context.ContactMessages.Where(x => x.DeletedAt != null).ToList();
+            return View(deletedMessages);
         }
 
         [HttpPost]
@@ -32,12 +39,43 @@ namespace App.Web.Mvc.Areas.Admin.Controllers
                 return NotFound();
             }
 
-            _context.ContactMessages.Remove(message);
+            message.DeletedAt = DateTime.UtcNow;
             _context.SaveChanges();
 
             return RedirectToAction("Index");
         }
 
+        [HttpPost]
+        public IActionResult Restore(int id)
+        {
+            var message = _context.ContactMessages.Find(id);
+
+            if (message == null)
+            {
+                return NotFound();
+            }
+
+            message.DeletedAt = null;
+            _context.SaveChanges();
+
+            return RedirectToAction("DeletedMessages");
+        }
+
+        [HttpPost]
+        public IActionResult PermanentlyDelete(int id)
+        {
+            var message = _context.ContactMessages.Find(id);
+
+            if (message == null)
+            {
+                return NotFound();
+            }
+
+            _context.ContactMessages.Remove(message);
+            _context.SaveChanges();
+
+            return RedirectToAction("DeletedMessages");
+        }
 
     }
 }
